Reorganize tests under task domain
